pager의 item이 선택됬을때 처리 stategy로 이동
